+++ src/v2019C1/extensions/history/query/get-sorted-hitem.test.ts
@@ -6,7 +6,7 @@ import { ALL_XMLNS_NAMESPACES, CUSTOM_RECORD_ID_ATTRIBUTE, runSclTestCases } fro
 
 import type { SclTest } from '@/v2019C1/test'
 
-type TestCase = SclTest.BaseTestCase & {
+type TestCase = SclTest.BaseXmlTestCase & {
 	expectedOrder: { version: string; revision: string }[]
 }